@@ -2,7 +2,7 @@
   "name": "react-tech-commerce.js-site",
   "version": "1.0.0",
   "description": "An eCommerce site which uses Commerce.js to create the store and the Stripe gateway to allow customers checkout. This is in sandbox mode.",
-  "main": "index.js",
+  "main": "index.tsx",
   "private": true,
   "scripts": {
     "start-server": "cross-env SERVE=true webpack serve",
@@ -23,38 +23,46 @@
   },
   "homepage": "https://github.com/asbhogal/React-Tech-Commerce.Js-Site",
   "devDependencies": {
-    "@babel/core": "^7.20.2",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.9",
-    "babel-loader": "^9.1.0",
+    "@babel/core": "^7.23.6",
+    "@babel/preset-env": "^7.23.6",
+    "@babel/preset-react": "^7.23.3",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.11",
+    "@types/node": "^20.10.5",
+    "@types/react": "^18.2.45",
+    "@types/react-dom": "^18.2.18",
+    "babel-loader": "^9.1.3",
     "clean-webpack-plugin": "^4.0.0",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.7.2",
+    "css-loader": "^6.8.1",
     "dotenv-webpack": "^8.0.1",
-    "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.6.1",
-    "postcss": "^8.4.19",
-    "postcss-loader": "^7.0.1",
-    "postcss-preset-env": "^7.8.3",
+    "html-webpack-plugin": "^5.6.0",
+    "mini-css-extract-plugin": "^2.7.6",
+    "postcss": "^8.4.32",
+    "postcss-loader": "^7.3.3",
+    "postcss-preset-env": "^9.3.0",
     "process": "^0.11.10",
     "react-refresh": "^0.14.0",
-    "sass": "^1.56.1",
-    "sass-loader": "^13.2.0",
-    "webpack": "^5.75.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.11.1"
+    "sass": "^1.69.5",
+    "sass-loader": "^13.3.3",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.3.3",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   },
   "dependencies": {
     "@chec/commerce.js": "^2.8.0",
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
-    "@stripe/react-stripe-js": "^1.16.1",
-    "@stripe/stripe-js": "^1.46.0",
-    "dompurify": "^2.4.3",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "latest",
+    "@mui/material": "latest",
+    "@stripe/react-stripe-js": "^2.4.0",
+    "@stripe/stripe-js": "^2.2.2",
+    "@types/dompurify": "^3.0.5",
+    "dompurify": "^3.0.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.41.3",
-    "react-router-dom": "^6.6.1"
+    "react-hook-form": "^7.49.2",
+    "react-router-dom": "^6.21.1"
   }
 }
