@@ -1,4 +1,4 @@
-import { TextField, Grid } from "@material-ui/core";
+import { TextField, Grid } from "@mui/material";
 import { useFormContext, Controller } from "react-hook-form";
 
 const FormInput = ({ name, label }) => {
