@@ -1,4 +1,4 @@
-import { Typography, Button, Divider } from "@material-ui/core";
+import { Typography, Button, Divider } from "@mui/material";
 
 import {
   Elements,
