@@ -1,22 +1,26 @@
 import React from "react";
-import { Grid } from "@material-ui/core";
+import { Box, Container, Grid } from "@mui/material";
 import Product from "./Product/Product";
-import useStyles from "./styles";
+import styles from "./styles";
 
-const Products = ({ products, onAddToCart }) => {
-  const classes = useStyles();
-
+const Products = ({
+  products,
+  onAddToCart,
+}: {
+  products: any;
+  onAddToCart: any;
+}) => {
   return (
-    <main className={classes.content}>
-      <div className={classes.toolbar} />
+    <Container sx={styles.content}>
+      <Box sx={styles.toolbar} />
       <Grid container justifyContent="center" spacing={4}>
-        {products.map((product) => (
+        {products.map((product: any) => (
           <Grid item key={product.id} xs={12} sm={6} md={4} lg={3}>
             <Product product={product} onAddToCart={onAddToCart} />
           </Grid>
         ))}
       </Grid>
-    </main>
+    </Container>
   );
 };
 
